Fixing issues with jpg endpoint
JPG endpoint used to initialize a new camera instance and take a snapshot. On some chips, this would cause the driver to error out and the entire camera to become unresponsive until a reboot. This change utilizes the frame held in the buffer and delivers it instead to prevent any errors.